pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.7.2"
numpy = "^1.25.2"
yfinance = "^0.2.28"
pandas = "^2.0.3"
scikit-learn = "^1.3.0"
xgboost = "^1.7.6"
tqdm = "^4.66.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
